use generic useStockGraphQL
